requirements.txt
streamlit
pandas
plotly
gspread
google-auth
wordcloud
matplotlib
scikit-learn
streamlit-plotly-events

#Update
